Restructuring page content layout related components.
